backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
cryptography==3.4.7
PyJWT==2.1.0